requirements.txt
matplotlib~=3.1.3
numpy~=1.18.1
imageio~=2.8.0
scipy~=1.4.1
matplot~=0.1.9
pandas~=1.0.3
plotly~=4.8.1
pillow~=7.1.2